rename local utilities module and remove shared utils
